+++ src/utils.js
@@ -1,9 +1,5 @@
 const log = function(){
-    const con = window.console;
-    
-    //if(COLOR.debug){
-        con.log.apply(null, Array.from( arguments ) );
-    //}
+    console.log.apply(null, Array.from( arguments ) );
 };
 const mod = function(x, m) {
     return (x % m + m) % m;